Implementing Book Repository

+++ composeApp/src/commonMain/kotlin/com/plcoding/bookpedia/book/domain/Book.kt
@@ -9,7 +9,7 @@ data class Book(
     val languages: List<String>,
     val firstPublishYear: String?,
     val averageRating: Double,
-    val ratingsCount: Int,
+    val ratingsCount: Double,
     val numPages: Int?,
     val nunEditions: Int
 )
